fuzzing: Refactor `TableOps` fuzz generator to allow GC with refs on the stack
This makes the generator more similar to `wasm-smith` where it is keeping track
of what is on the stack and making choices about what instructions are valid to
generate given the current stack state. This allows the generator to emit GC
calls while there are live refs on the stack, which it didn't do before.

Fixes #3917