fuzzing: Don't add `u64::MAX` fuel by default
Add a more reasonable number by default instead. Adding max fuel is problematic
in particular because the way that fuel gets counted, once we have max fuel it
is easy to add/remove fuel in ways that would overflow, but which we check and
have some default behavior that isn't always exactly what we want, and means
that our fuel count gets thrown off afterwards.