runtime: Don't GC when we can just allocate a bump chunk instead
When entering Wasm and passing `externref` arguments to the Wasm function, we
check that the `VMExternRefActivationsTable`'s bump chunk can hold all the
arguments. If it can't we would do a GC and then afterwards we would know that
the bump chunk was empty and could therefore hold the `externref`
arguments. However, we recently made the bump chunk lazily allocated. So the
first time we call into Wasm and pass `externref`s, we would not have any
capacity in the bump chunk, and then we would do a GC (which would then force
the lazy allocation of the bump chunk). But there is no need to actually GC and
walk the stack and all that in this situation, we just need to allocate the bump
chunk. This commit switches to that behavior.