Cranelift: Handle dynamic bounds checks uniformly
Unify the two code paths for dynamic memory bounds checking with vs without
Spectre. Now the `SpectreOobComparison` struct has been renamed to
`OobComparison` and gained a `spectre: bool` field. We look at this field to
determine whether to emit a regular `select` or a `select_spectre_guard`.

This also fixes the implementation of `heap_addr` with dynamic memories and
without Spectre mitigations enabled to match the documented semantics. It was
previously emitting the out-of-bounds trap itself, rather than returning the
`NULL` address or another address that is guaranteed to trap when accessed.