Handle all `decommit`s/`madvise(DONTNEED)`s homogenously (#8587)
* Don't differentiate between decommitting table vs stack pages

All implementations are the same either way, and by treating them differently,
it makes it difficult to prototype integration with batching syscalls that need
to treat them homogenously.

Co-Authored-By: Jamey Sharp <jsharp@fastly.com>

* Remove the `madvise_dontneed` function from the system virtual memory interface

This was already implemented by the existing `decommit_pages` function, we just
need to determine what the behavior is for the left over mapping. Specifically,
whether it is zeroed or restored to the original mapping (e.g. a CoW image).

This allows us to handle `decommit_pages` and what previously were calls to
`madvise_dontneed` homogenously, which enables us to prototype new system calls
that batch decommits together and therefore must treat them all the same.

Co-Authored-By: Jamey Sharp <jsharp@fastly.com>