Wasmtime(pooling allocator): Batch decommits (#8590)
This introduces a `DecommitQueue` for batching decommits together in the pooling
allocator:

* Deallocating a memory/table/stack enqueues their associated regions of memory
  for decommit; it no longer immediately returns the associated slot to the
  pool's free list. If the queue's length has reached the configured batch size,
  then we flush the queue by running all the decommits, and finally returning
  the memory/table/stack slots to their respective pools and free lists.

* Additionally, if allocating a new memory/table/stack fails because the free
  list is empty (aka we've reached the max concurrently-allocated limit for this
  entity) then we fall back to a slow path before propagating the error. This
  slow path flushes the decommit queue and then retries allocation, hoping that
  the queue flush reclaimed slots and made them available for this fallback
  allocation attempt. This involved defining a new `PoolConcurrencyLimitError`
  to match on, which is also exposed in the public embedder API.

It is also worth noting that we *always* use this new decommit queue now. To
keep the existing behavior, where e.g. a memory's decommits happen immediately
on deallocation, you can use a batch size of one. This effectively disables
queueing, forcing all decommits to be flushed immediately.

The default decommit batch size is one.

This commit, with batch size of one, consistently gives me an increase on
`wasmtime serve`'s requests-per-second versus its parent commit, as measured by
`benches/wasmtime-serve-rps.sh`. I get ~39K RPS on this commit compared to ~35K
RPS on the parent commit. This is quite puzzling to me. I was expecting no
change, and hoping there wouldn't be a regression. I was not expecting a speed
up. I cannot explain this result at this time.

prtest:full

Co-authored-by: Jamey Sharp <jsharp@fastly.com>